CHAT-45: Extract socket chat code to multiple files
