+++ chatmeleon-app/app/conversations/[conversationId]/example/socket-real-time-chat/types.d.ts
@@ -1,0 +1,7 @@
+export type Message = {
+  id: string;
+  body?: string;
+  image?: string;
+  createdAt: Date;
+  senderId: string;
+};
